Move constants file to support directory
