+++ exercises/00-init/cypress/integration/home_page_spec.ts
@@ -1,4 +1,4 @@
-import { USER_LIST } from './utils/constants'
+import { USER_LIST } from '../support/constants'
 
 describe('The Home Page', () => {
   it('successfully loads', () => {
