+++ exercises/00-init/cypress/support/constants.ts
