+++ exercises/01-react-stateful/cypress/support/constants.ts
